Organise the folders in commands
