+++ src/main/java/seedu/address/logic/commands/AddCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands.delete;
+package seedu.address.logic.commands;
 
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
@@ -6,10 +6,10 @@ import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 
 /**
- * Deletes the specified data from the address book.
+ * Adds the specified data to the address book.
  */
-public abstract class DeleteCommand extends Command {
-    public static final String COMMAND_WORD = "delete";
+public abstract class AddCommand extends Command {
+    public static final String COMMAND_WORD = "add";
 
     @Override
     public abstract CommandResult execute(Model model) throws CommandException;
