+++ src/main/java/seedu/address/logic/commands/applicant/DeleteApplicantCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands.delete;
+package seedu.address.logic.commands.applicant;
 
 import static java.util.Objects.requireNonNull;
 
@@ -8,6 +8,7 @@ import seedu.address.commons.core.DataType;
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.DeleteCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.applicant.Applicant;
