+++ src/main/java/seedu/address/logic/commands/position/AddPositionCommand.java
@@ -8,7 +8,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_REQUIREMENT;
 
 import seedu.address.commons.core.DataType;
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.logic.commands.add.AddCommand;
+import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.position.Position;
