+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -11,11 +11,11 @@ import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.EditCommand;
 import seedu.address.logic.commands.ExitCommand;
 import seedu.address.logic.commands.FindCommand;
-import seedu.address.logic.commands.add.AddCommand;
-import seedu.address.logic.commands.delete.DeleteCommand;
+import seedu.address.logic.commands.AddCommand;
+import seedu.address.logic.commands.DeleteCommand;
 import seedu.address.logic.commands.help.HelpCommand;
-import seedu.address.logic.commands.list.ListApplicantCommand;
-import seedu.address.logic.commands.list.ListCommand;
+import seedu.address.logic.commands.applicant.ListApplicantCommand;
+import seedu.address.logic.commands.ListCommand;
 import seedu.address.logic.parser.applicants.EditApplicantCommandParser;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.logic.parser.help.HelpCommandParser;
