+++ src/main/java/seedu/address/logic/parser/applicants/AddApplicantCommandParser.java
@@ -12,7 +12,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
 import java.util.Set;
 import java.util.stream.Stream;
 
-import seedu.address.logic.commands.add.AddApplicantCommand;
+import seedu.address.logic.commands.applicant.AddApplicantCommand;
 import seedu.address.logic.parser.ArgumentMultimap;
 import seedu.address.logic.parser.ArgumentTokenizer;
 import seedu.address.logic.parser.Parser;
