+++ src/main/java/seedu/address/logic/parser/applicants/DeleteApplicantCommandParser.java
@@ -3,7 +3,7 @@ package seedu.address.logic.parser.applicants;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.delete.DeleteApplicantCommand;
+import seedu.address.logic.commands.applicant.DeleteApplicantCommand;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
