+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -24,7 +24,7 @@ import org.junit.jupiter.api.io.TempDir;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.add.AddApplicantCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.logic.commands.list.ListApplicantCommand;
+import seedu.address.logic.commands.applicant.ListApplicantCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
