+++ src/test/java/seedu/address/logic/commands/AddApplicantCommandIntegrationTest.java
@@ -7,7 +7,7 @@ import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.add.AddApplicantCommand;
+import seedu.address.logic.commands.applicant.AddApplicantCommand;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
