+++ src/test/java/seedu/address/logic/commands/DeleteApplicantCommandTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.delete.DeleteApplicantCommand;
+import seedu.address.logic.commands.applicant.DeleteApplicantCommand;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
