+++ src/test/java/seedu/address/logic/commands/ListApplicantCommandTest.java
@@ -8,7 +8,7 @@ import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.list.ListApplicantCommand;
+import seedu.address.logic.commands.applicant.ListApplicantCommand;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
