+++ src/test/java/seedu/address/logic/parser/AddressBookParserTest.java
@@ -20,9 +20,9 @@ import seedu.address.logic.commands.EditCommand.EditPersonDescriptor;
 import seedu.address.logic.commands.ExitCommand;
 import seedu.address.logic.commands.FindCommand;
 import seedu.address.logic.commands.add.AddApplicantCommand;
-import seedu.address.logic.commands.delete.DeleteApplicantCommand;
+import seedu.address.logic.commands.applicant.DeleteApplicantCommand;
 import seedu.address.logic.commands.help.HelpCommand;
-import seedu.address.logic.commands.list.ListApplicantCommand;
+import seedu.address.logic.commands.applicant.ListApplicantCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.applicant.Applicant;
 import seedu.address.model.applicant.NameContainsKeywordsPredicate;
