+++ src/test/java/seedu/address/logic/parser/DeleteApplicantCommandParserTest.java
@@ -8,7 +8,7 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.delete.DeleteApplicantCommand;
+import seedu.address.logic.commands.applicant.DeleteApplicantCommand;
 
 /**
  * As we are only doing white-box testing, our test cases do not cover path variations
